+++ src/Providers/RouteServiceProvider.php
@@ -37,9 +37,11 @@ class RouteServiceProvider extends ServiceProvider
      */
     protected function mapWebRoutes()
     {
-        Route::middleware('web')
-             ->namespace($this->namespace)
-             ->group(base_path('routes/web.php'));
+        if (is_file($routes = base_path('routes/web.php'))) {
+            Route::middleware('web')
+                 ->namespace($this->namespace)
+                 ->group($routes);
+        }
     }
 
     /**
@@ -51,8 +53,10 @@ class RouteServiceProvider extends ServiceProvider
      */
     protected function mapApiRoutes()
     {
-        Route::middleware('api')
-             ->namespace($this->namespace)
-             ->group(base_path('routes/api.php'));
+        if (is_file($routes = base_path('routes/api.php'))) {
+            Route::middleware('api')
+                ->namespace($this->namespace)
+                ->group($routes);
+        }
     }
 }