+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+- Use correct namespace for GraylogLoggerFactory
 
 ## [0.2.6] - 2020-04-30